refactor: implement centralized app state management and enhance header functionality
- Created `AppContext` and `AppProvider` for managing global state, including `showRequirements` and `isSidePanelOpen`.
  - Provides a cleaner and reusable way to manage application-wide states.
  - Prevents redundant state management across components.
- Refactored `HeaderWrapper` to integrate `AppContext` for toggling between requirements and article views.
  - Added `handleHomeClick` to reset state when navigating to the homepage.
  - Enhanced navigation logic for consistent user experience.
- Updated `Header` component to accept `onHomeClick` for better integration with navigation logic.
- Adjusted `Page` and `MainContent` components to use `useApp` for accessing global state instead of maintaining local states.
- Improved maintainability and readability of state management logic throughout the app.
